requirements.txt
oracledb==3.1.1
python-dotenv==1.1.0